# openapi.yml
openapi: 3.1.0
info:
  title: EDData API
  description: |
    REST API for accessing Elite Dangerous game data collected from EDDN (Elite Dangerous Data Network).
    
    The API provides comprehensive data about:
    - Over 150 million star systems with detailed location data
    - Millions of trade orders for commodities and markets
    - Hundreds of thousands of stations, ports, and fleet carriers
    - Real-time updates from the Elite Dangerous community
    
    **Related Services:**
    - [EDData Collector](https://github.com/EDDataAPI/eddata-collector) - Data collection from EDDN
    - [EDData Web](https://github.com/EDDataAPI/eddata-web) - Web interface
    - [EDData Auth](https://github.com/EDDataAPI/eddata-auth) - Authentication service
  version: 1.0.0
  contact:
    name: EDData API Support
    url: https://github.com/EDDataAPI/eddata-api
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0
servers:
  - url: https://api.eddata.dev
    description: Production API
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/v2/commodity/name/{commodityName}:
    get:
      tags:
        - Commodities
      summary: Get commodity information
      description: Returns summary report for a specific commodity
      operationId: getCommodity
      parameters:
        - $ref: '#/components/parameters/CommodityName'
      responses:
        '200':
          description: Commodity information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommodityReport'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/commodity/name/{commodityName}/imports:
    get:
      tags:
        - Commodities
      summary: Get commodity importers
      description: Returns places importing a commodity (where you can sell), ordered by highest price. Returns best 100 matching results.
      operationId: getCommodityImporters
      parameters:
        - $ref: '#/components/parameters/CommodityName'
        - $ref: '#/components/parameters/MinVolume'
        - $ref: '#/components/parameters/MinPrice'
        - $ref: '#/components/parameters/FleetCarriers'
        - $ref: '#/components/parameters/MaxDaysAgo'
      responses:
        '200':
          description: List of importers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommodityOrder'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/commodity/name/{commodityName}/exports:
    get:
      tags:
        - Commodities
      summary: Get commodity exporters
      description: Returns places exporting a commodity (where you can buy), ordered by lowest price. Returns best 100 matching results.
      operationId: getCommodityExporters
      parameters:
        - $ref: '#/components/parameters/CommodityName'
        - $ref: '#/components/parameters/MinVolume'
        - $ref: '#/components/parameters/MaxPrice'
        - $ref: '#/components/parameters/FleetCarriers'
        - $ref: '#/components/parameters/MaxDaysAgo'
      responses:
        '200':
          description: List of exporters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommodityOrder'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v2/market/{marketId}/commodity/name/{commodityName}:
    get:
      tags:
        - Markets
      summary: Get commodity data for specific market
      description: Returns commodity information for a specific market (useful for rare goods)
      operationId: getMarketCommodity
      parameters:
        - name: marketId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 128106744
        - $ref: '#/components/parameters/CommodityName'
      responses:
        '200':
          description: Commodity data for market
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommodityOrder'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v2/search/system/name/{searchTerm}:
    get:
      tags:
        - Search
      summary: Search systems by name
      description: Search for star systems by name (partial match supported)
      operationId: searchSystems
      parameters:
        - name: searchTerm
          in: path
          required: true
          schema:
            type: string
          example: "Sol"
      responses:
        '200':
          description: List of matching systems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/System'
components:
  parameters:
...
    CommodityName:
      name: commodityName
      in: path
      required: true
      schema:
        type: string
      example: "gold"
      description: Commodity name (lowercase, no special characters)
    MinVolume:
      name: minVolume
      in: query
      schema:
        type: integer
        default: 1
      description: Minimum volume/stock
    MinPrice:
      name: minPrice
      in: query
      schema:
        type: integer
        default: 1
      description: Minimum price for imports
    MaxPrice:
      name: maxPrice
      in: query
      schema:
        type: integer
      description: Maximum price for exports
    FleetCarriers:
      name: fleetCarriers
      in: query
      schema:
        type: boolean
      description: |
        Filter by fleet carriers:
        - true: Only fleet carriers
        - false: Exclude fleet carriers
        - null/omitted: Include all
    MaxDaysAgo:
      name: maxDaysAgo
      in: query
      schema:
        type: integer
        default: 30
        maximum: 90
      description: Maximum age of data in days
...
  schemas:
    System:
      type: object
      properties:
        systemName:
          type: string
          example: "Sol"
        systemAddress:
          type: integer
          format: int64
          example: 10477373803
        systemX:
          type: number
          format: double
          example: 0.0
        systemY:
          type: number
          format: double
          example: 0.0
        systemZ:
          type: number
          format: double
          example: 0.0
        updatedAt:
          type: string
          format: date-time
...
    CommodityReport:
      type: object
      properties:
        commodityName:
          type: string
          example: "gold"
        maxBuyPrice:
          type: integer
          example: 59797
        minBuyPrice:
          type: integer
          example: 3979
        avgBuyPrice:
          type: integer
          example: 44441
        totalStock:
          type: integer
          format: int64
          example: 73016533
        maxSellPrice:
          type: integer
          example: 70761
        minSellPrice:
          type: integer
          example: 3978
        avgSellPrice:
          type: integer
          example: 48259
        totalDemand:
          type: integer
          format: int64
          example: 1899662825
    CommodityOrder:
      type: object
      properties:
        commodityName:
          type: string
        marketId:
          type: integer
          format: int64
        stationName:
          type: string
        stationType:
          type: string
        distanceToArrival:
          type: number
          format: double
        maxLandingPadSize:
          type: string
          enum: [S, M, L]
        bodyId:
          type: integer
        bodyName:
          type: string
        systemAddress:
          type: integer
          format: int64
        systemName:
          type: string
        systemX:
          type: number
          format: double
        systemY:
          type: number
          format: double
        systemZ:
          type: number
          format: double
        buyPrice:
          type: integer
        demand:
          type: integer
        demandBracket:
          type: integer
          minimum: 0
          maximum: 3
        meanPrice:
          type: integer
        sellPrice:
          type: integer
        stock:
          type: integer
        stockBracket:
          type: integer
          minimum: 0
          maximum: 3
        updatedAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not Found"
            message: "The requested resource was not found"
  securitySchemes: {}
security: []